Combine `HandleApiRelations` and `HandleApiBrowsers` traits [MA-124]
