+++ app/Repositories/Behaviors/HandleApiBrowsers.php
@@ -2,15 +2,16 @@
 
 namespace App\Repositories\Behaviors;
 
-use App\Models\Behaviors\HasMedias;
+use A17\Twill\Models\RelatedItem;
 use App\Models\ApiRelation;
+use App\Libraries\Api\Models\BaseApiModel;
 use App\Helpers\UrlHelpers;
 use DamsImageService;
-use Illuminate\Database\Eloquent\Model as EloquentModel;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\Relations\HasOne;
 use Illuminate\Database\Eloquent\Relations\MorphTo;
+use Illuminate\Database\Eloquent\Relations\MorphToMany;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Str;
 
@@ -50,13 +51,23 @@ trait HandleApiBrowsers
     public function afterSaveHandleApiBrowsers($object, $fields)
     {
         foreach ($this->getApiBrowsers() as $browser) {
-            $this->updateBrowserApi(
-                $object,
-                $fields,
-                $browser['relation'],
-                $browser['positionAttribute'],
-                $browser['browserName']
-            );
+            if ($browser['isApiRelation']) {
+                $this->updateBrowserApiRelated(
+                    $object,
+                    $fields,
+                    $browser['relation'],
+                    $browser['positionAttribute']
+                );
+            }
+            else {
+                $this->updateApiBrowser(
+                    $object,
+                    $fields,
+                    $browser['relation'],
+                    $browser['positionAttribute'],
+                    $browser['browserName']
+                );
+            }
         }
     }
 
@@ -70,9 +81,10 @@ trait HandleApiBrowsers
         foreach ($this->getApiBrowsers() as $browser) {
             $relation = $browser['relation'];
             if (collect($object->$relation)->isNotEmpty()) {
-                $fields['browsers'][$browser['browserName']] = $this->getFormFieldsForBrowserApi(
+                $fields['browsers'][$browser['browserName']] = $this->getFormFieldsForApiBrowser(
                     $object,
                     $relation,
+                    $browser['model'],
                     $browser['routePrefix'],
                     $browser['titleKey'],
                     $browser['moduleName']
@@ -92,7 +104,7 @@ trait HandleApiBrowsers
      * @param array $pivotAttributes
      * @return void
      */
-    public function updateBrowserApi(
+    public function updateApiBrowser(
         $object,
         $fields,
         $relationship,
@@ -131,13 +143,13 @@ trait HandleApiBrowsers
             $object->$relationship() instanceof HasOne ||
             $object->$relationship() instanceof HasMany
         ) {
-            $this->updateBelongsToInverseBrowserApi($object, $relationship, $relatedElements);
+            $this->updateBelongsToInverseApiBrowser($object, $relationship, $relatedElements);
         } else {
             $object->$relationship()->sync($relatedElementsWithPosition);
         }
     }
 
-    private function updateBelongsToInverseBrowserApi($object, $relationship, $updatedElements)
+    private function updateBelongsToInverseApiBrowser($object, $relationship, $updatedElements)
     {
         $foreignKey = $object->$relationship()->getForeignKeyName();
         $relatedModel = $object->$relationship()->getRelated();
@@ -165,7 +177,7 @@ trait HandleApiBrowsers
      */
     public function updateOrderedBelongsTomany($object, $fields, $relationship, $positionAttribute = 'position')
     {
-        $this->updateBrowserApi($object, $fields, $relationship, $positionAttribute);
+        $this->updateApiBrowser($object, $fields, $relationship, $positionAttribute);
     }
 
     /**
@@ -187,7 +199,7 @@ trait HandleApiBrowsers
      * @param string|null $moduleName
      * @return array
      */
-    public function getFormFieldsForBrowserApi(
+    public function getFormFieldsForApiBrowser(
         $object,
         $relation,
         $apiModel,
@@ -195,36 +207,42 @@ trait HandleApiBrowsers
         $titleKey = 'title',
         $moduleName = null
     ) {
-        if ($object->{$relation}->isEmpty()) {
+        if (!$object->{$relation} || $object->{$relation}->isEmpty()) {
             return [];
         }
 
-        $fields = $this->getRelatedElementsAsCollection($object, $relation);
+        $related = $this->getRelatedElementsAsCollection($object, $relation);
 
-        $isMorphTo = method_exists($object, $relation) && $object->$relation() instanceof MorphTo;
-
-        if ($fields->isNotEmpty()) {
-            if ($fields->first() instanceof ApiRelation) {
-                $ids = $object->{$relation}->pluck('datahub_id')->toArray();
-                $fields = $apiModel::query()->ids($ids)->get();
-            }
-            return $fields->map(
-                function ($relatedElement) use ($titleKey, $routePrefix, $relation, $moduleName, $isMorphTo) {
-                    if ($isMorphTo && !$moduleName) {
-                        // @todo: Maybe there is an existing helper for this?
-                        $moduleName = Str::plural(Arr::last(explode('\\', get_class($relatedElement))));
-                    }
-
-                    return [
-                        'id' => $relatedElement->id,
-                        'name' => $relatedElement->titleInBrowser ?? $relatedElement->$titleKey,
-                        'endpointType' => $relatedElement->getMorphClass(),
-                    ] + $this->getDataEditAndThumbnailAttribute($relatedElement, $routePrefix, $relation, $moduleName);
-                }
-            )->values()->toArray();
+        if ($related->isEmpty()) {
+            return [];
         }
 
-        return [];
+        $isMorphTo = method_exists($object, $relation) && $object->$relation() instanceof MorphTo;
+        if ($related->first() instanceof ApiRelation) {
+            $ids = $object->{$relation}->pluck('datahub_id')->toArray();
+            $related = $apiModel::query()->ids($ids)->get();
+        }
+        if ($related->first() instanceof BaseApiModel) {
+            $browserData = $related->map(function ($relatedElement) use ($relation, $titleKey, $moduleName) {
+                return $this->buildBrowserData($relation, $relatedElement, $titleKey, $moduleName);
+            });
+        }
+        else {
+            // Get all datahub_id's
+            $ids = $related->pluck('datahub_id')->toArray();
+            // Use those to load API records
+            $apiElements = $apiModel::query()->ids($ids)->get();
+            // Find locally selected objects
+            $localApiMapping = $related->filter(function ($relatedElement) use ($apiElements) {
+                return $apiElements->where('id', $relatedElement->datahub_id)->first();
+            });
+            $browserData = $localApiMapping->map(function ($relatedElement) use ($titleKey, $routePrefix, $relation, $moduleName, $apiElements) {
+                // Get the API elements and use them to build the browser elements
+                $apiElement = $apiElements->where('id', $relatedElement->datahub_id)->first();
+                return $this->buildBrowserData($relation, $apiElement, $titleKey, $moduleName);
+            });
+        }
+        return $browserData->values()->toArray();
     }
 
     /**
@@ -232,7 +250,7 @@ trait HandleApiBrowsers
      * @param string $relation
      * @return array
      */
-    public function getFormFieldsForRelatedBrowserApi($object, $relation, $titleKey = 'title')
+    public function getFormFieldsForRelatedApiBrowser($object, $relation, $titleKey = 'title')
     {
         return $object->getRelated($relation)->map(function ($relatedElement) use ($titleKey, $relation) {
             return ($relatedElement != null) ? [
@@ -253,7 +271,7 @@ trait HandleApiBrowsers
         if (method_exists($relatedElement, 'hasAugmentedModel') && $relatedElement->hasAugmentedModel() && $relatedElement->getAugmentedModel()) {
             $data['edit'] = $relatedElement->adminEditUrl ?? moduleRoute($moduleName ?? $relation, $routePrefix ?? '', 'edit', [$relatedElement->getAugmentedModel()->id]);
 
-            if (classHasTrait($relatedElement->getAugmentedModel(), HasMedias::class)) {
+            if (classHasTrait($relatedElement->getAugmentedModel(), \App\Models\Behaviors\HasMedias::class)) {
                 $data['thumbnail'] = $relatedElement->getAugmentedModel()->defaultCmsImage(['w' => 100, 'h' => 100, 'fit' => 'crop']);
             }
         } else {
@@ -297,6 +315,7 @@ trait HandleApiBrowsers
                 'model' => empty($browser['model']) ? 'App\\Models\\Api\\' . $this->inferModelFromModuleName($moduleName) : $browser['model'],
                 'positionAttribute' => empty($browser['positionAttribute']) ? 'position' : $browser['positionAttribute'],
                 'browserName' => $browserName,
+                'isApiRelation' => isset($browser['isApiRelation']) && $browser['isApiRelation'] == true,
             ];
         })->values();
     }
@@ -328,7 +347,168 @@ trait HandleApiBrowsers
     private function getRelatedElementsAsCollection($object, $relation)
     {
         return collect(
-            $object->$relation instanceof EloquentModel ? [$object->$relation] : $object->$relation
+            $object->$relation instanceof BaseApiModel ? [$object->$relation] : $object->$relation
         );
     }
+
+    /**
+     * The same as the normal ordered update with the difference that this one adds a relation to the pivot
+     * and it creates new models per each new relation as we don't have both ends of the polymorphic relation
+     * This is done this way so we can reuse the same functions and table for all API browsers.
+     */
+    public function updateBrowserApiRelated($object, $fields, $relationship, $positionAttribute = 'position')
+    {
+        $relatedElementsWithPosition = [];
+
+        $fieldsHasElements = isset($fields['browsers'][$relationship]) && !empty($fields['browsers'][$relationship]);
+        $relatedElements = $fieldsHasElements ? $fields['browsers'][$relationship] : [];
+
+        // If we don't have an element to save the datahub_id, let's create one
+        $relatedElements = array_map(function ($element) {
+            return ApiRelation::firstOrCreate(['datahub_id' => $element['id']]);
+        }, $relatedElements);
+
+        $position = 1;
+
+        foreach ($relatedElements as $relatedElement) {
+            $relatedElementsWithPosition[$relatedElement['id']] = [
+                // Add the relationship to the pivot, this way we can use this browser several times per model
+                'relation' => $relationship,
+                $positionAttribute => $position++
+            ];
+        }
+        $object->{$relationship}()->detach($object->{$relationship}->pluck('id'));
+        $object->{$relationship}()->attach($relatedElementsWithPosition);
+    }
+
+    public function updateMultiBrowserApiRelated($object, $fields, $relationship, $typeUsesApi)
+    {
+        // WEB-2272: check if we dont leave some stale data in database by not deleting apiElements
+        // Remove all associations
+        // $object->apiElements()->detach();
+
+        $relatedElementsWithPosition = [];
+
+        $fieldsHasElements = isset($fields['browsers'][$relationship]) && !empty($fields['browsers'][$relationship]);
+        $relatedElements = $fieldsHasElements ? $fields['browsers'][$relationship] : [];
+        // If we don't have an element to save the datahub_id, let's create one
+        $relatedElements = array_map(function ($element) use ($typeUsesApi) {
+            if ($typeUsesApi[$element['endpointType']]) {
+                $apiItem = ApiRelation::firstOrCreate(['datahub_id' => $element['id']]);
+                $apiItem->endpointType = $element['endpointType'];
+
+                return $apiItem;
+            }
+
+            return $element;
+        }, $relatedElements);
+
+        RelatedItem::where([
+            'browser_name' => $relationship,
+            'subject_id' => $object->getKey(),
+            'subject_type' => $object->getMorphClass(),
+        ])->delete();
+
+        $position = 1;
+        collect($relatedElements)->each(function ($values) use ($relationship, &$position, $object) {
+            RelatedItem::create([
+                'subject_id' => $object->getKey(),
+                'subject_type' => $object->getMorphClass(),
+                'related_id' => $values['id'],
+                'related_type' => $values['endpointType'],
+                'browser_name' => $relationship,
+                'position' => $position,
+            ]);
+            $position++;
+        });
+    }
+
+    protected function buildBrowserData($relation, $apiElement, $titleKey, $moduleName)
+    {
+        $data = [];
+        // If it contains an augmented model create an edit link
+        if ($apiElement->hasAugmentedModel() && $apiElement->getAugmentedModel()) {
+            $data['edit'] = moduleRoute($moduleName ?? $relation, $routePrefix ?? '', 'edit', [$apiElement->getAugmentedModel()->id]);
+
+            if (classHasTrait($apiElement->getAugmentedModel(), \App\Models\Behaviors\HasMedias::class)) {
+                $data['thumbnail'] = $apiElement->getAugmentedModel()->defaultCmsImage(['w' => 100, 'h' => 100]);
+            }
+        } else {
+            // WEB-1187: This is reached after page refresh, if the model hasn't been augmented
+            if (UrlHelpers::moduleRouteExists($moduleName ?? $relation, $routePrefix ?? '', 'augment')) {
+                $data['edit'] = moduleRoute($moduleName ?? $relation, $routePrefix ?? '', 'augment', [$apiElement->id]);
+            }
+
+            if ($apiElement->hasAttribute('image_id')) {
+                $data['thumbnail'] = DamsImageService::getTransparentFallbackUrl(['w' => 100, 'h' => 100]);
+            }
+        }
+
+        return [
+            'id' => $apiElement->id,
+            'name' => $apiElement->titleInBrowser ?? $apiElement->{$titleKey},
+        ] + $data;
+    }
+
+    public function getFormFieldsForMultiBrowserApi($object, $browser_name, $apiModelsDefinitions, $typeUsesApi)
+    {
+        $results = collect();
+
+        $typedFormFields = $object->relatedItems
+            ->where('browser_name', $browser_name)
+            ->groupBy('related_type')
+            ->map(function ($items, $type) use ($apiModelsDefinitions, $browser_name, $typeUsesApi) {
+                if ($typeUsesApi[$type]) {
+                    $apiElements = $this->getApiElements($items, $type, $apiModelsDefinitions);
+                    $localApiMapping = $this->getLocalApiMapping($items, $apiElements);
+                    $apiModelDefinition = $apiModelsDefinitions[$type];
+
+                    return $localApiMapping->map(function ($relatedElement) use ($apiModelDefinition, $apiElements) {
+                        $data = [];
+                        // Get the API elements and use them to build the browser elements
+                        $apiRelationElement = \App\Models\ApiRelation::where('id', $relatedElement->related_id)->first();
+                        $apiElement = $apiElements->where('id', $apiRelationElement->datahub_id)->first();
+
+                        // If it contains an augmented model create an edit link
+                        if ($apiElement->hasAugmentedModel() && $apiElement->getAugmentedModel()) {
+                            $data['edit'] = moduleRoute($apiModelDefinition['moduleName'], $apiModelDefinition['routePrefix'] ?? '', 'edit', [$apiElement->getAugmentedModel()->id]);
+
+                            if (classHasTrait($apiElement->getAugmentedModel(), \App\Models\Behaviors\HasMedias::class)) {
+                                $data['thumbnail'] = $apiElement->getAugmentedModel()->defaultCmsImage(['w' => 100, 'h' => 100]);
+                            }
+                        }
+                        // WEB-1187: Add augment route here!
+
+                        return [
+                            'id' => $apiElement->id,
+                            'name' => $apiElement->titleInBrowser ?? $apiElement->title,
+                            'endpointType' => $apiModelDefinition['moduleName'],
+                            'position' => $relatedElement->position
+                        ] + $data;
+                    })->values()->toArray();
+                } else {
+                    return $items->map(function ($relatedElement) {
+                        $element = $relatedElement->related;
+                        $elementPosition = $relatedElement->position;
+
+                        if ($element) {
+                            return [
+                                'id' => $element->id,
+                                'name' => $element->titleInBrowser ?? $element->title,
+                                'endpointType' => $element->getMorphClass(),
+                                'position' => $elementPosition,
+                                'edit' => $element->adminEditUrl,
+                            ] + ((classHasTrait($element, \A17\Twill\Models\Behaviors\HasMedias::class)) ? [
+                                'thumbnail' => $element->defaultCmsImage(['w' => 100, 'h' => 100]),
+                            ] : []);
+                        }
+                    });
+                }
+            });
+
+        return $typedFormFields->flatten(1)->sortBy(function ($browserItem, $key) {
+            return $browserItem['position'];
+        })->values()->toArray();
+    }
+
 }
